Now only does memory mapped build if you have more than 6GB of mem.

+++ _maked2g.sh
@@ -4,10 +4,14 @@ source paths
 export PATH=$PATH:$SCRIPTPATH
 FILENAME="CM4D2G-NUKE-`date +%Y%m%d`.zip"
 cd $REPOPATH
+MEM=`awk '/MemTotal/ {printf( "%.0f\n", $2 / 1024 )}' /proc/meminfo`
+if [ $MEM -ge 6000 ]
+then
 umount $REPOPATH/out
 rm -Rf $REPOPATH/out
 mkdir -p $REPOPATH/out
 mount -t tmpfs -o size=6G tmpfs $REPOPATH/out
+fi
 make clobber
 repo sync -j32
 export USE_CCHACHE=1
@@ -23,5 +27,9 @@ if [ -e $REPOPATH/out/target/product/droid2we/update-squished.zip ]
   else
     echo "SOME SHIT BE FUKT!"
 fi
+MEM=`awk '/MemTotal/ {printf( "%.0f\n", $2 / 1024 )}' /proc/meminfo`
+if [ $MEM -ge 6000 ]
+then
 umount $REPOPATH/out
 rm -Rf $REPOPATH/out
+fi
